refactor: HTTP 객체 만드는 Generator 객체 추가
